+++ packages/lex_gen/lib/src/services/fmt/lex_packages_generator.dart
@@ -4,12 +4,14 @@
 
 // Project imports:
 import '../../model/nsid.dart';
+import '../gen_context.dart';
 import '../object/lex_package.dart';
 import '../object/lex_service.dart';
 import '../object/lex_type.dart';
 import '../rule.dart' as rule;
 
 List<LexPackage> generateLexPackagesForService(
+  final GenContext ctx,
   final List<LexService> services,
 ) {
   final packages = <String, Set<String>>{};
@@ -20,6 +22,7 @@ List<LexPackage> generateLexPackagesForService(
         .putIfAbsent(key, () => <String>{})
         .add(
           rule.getLexObjectAbsolutePathForService(
+            ctx,
             service.lexiconId,
             service.getFileName(),
           ),
@@ -30,7 +33,7 @@ List<LexPackage> generateLexPackagesForService(
   for (final package in packages.entries) {
     final lexiconId = package.key;
 
-    final root = rule.getRootPackageName(lexiconId);
+    final root = rule.getRootPackageName(ctx, lexiconId);
     final name = rule.getPackageName(
       lexiconId.substring(0, lexiconId.length - 1),
     );
@@ -47,21 +50,30 @@ List<LexPackage> generateLexPackagesForService(
   return lexPackages;
 }
 
-List<LexPackage> generateLexPackages(final List<GeneratableType> types) {
+List<LexPackage> generateLexPackages(
+  final GenContext ctx,
+  final List<GeneratableType> types,
+) {
   final packages = <String, Set<String>>{};
   for (final type in types) {
     if (type.isShouldNotBeGenerated()) continue;
 
     packages
         .putIfAbsent(type.lexiconId, () => <String>{})
-        .add(rule.getLexObjectAbsolutePath(type.lexiconId, type.getFileName()));
+        .add(
+          rule.getLexObjectAbsolutePath(
+            ctx,
+            type.lexiconId,
+            type.getFileName(),
+          ),
+        );
   }
 
   final lexPackages = <LexPackage>[];
   for (final package in packages.entries) {
     final lexiconId = package.key;
 
-    final root = rule.getRootPackageName(lexiconId);
+    final root = rule.getRootPackageName(ctx, lexiconId);
     final name = rule.getPackageName(lexiconId);
 
     lexPackages.add(