+++ packages/lex_gen/lib/src/services/fmt/lex_xrpc_query_generator.dart
@@ -6,18 +6,21 @@
 import 'package:lexicon/lexicon.dart' as lex;
 
 // Project imports:
+import '../gen_context.dart';
 import '../object/lex_input.dart';
 import '../object/lex_output.dart';
 import '../rule.dart' as rule;
 import 'lex_property_generator.dart';
 
 (LexInput?, LexOutput?)? generateLexXrpcQuery(
+  final GenContext ctx,
   final lex.NSID lexiconId,
   final String defName,
   final lex.LexXrpcQuery query,
   final List<String> mainVariants,
 ) {
   return _LexXrpcQueryGenerator(
+    ctx,
     lexiconId,
     defName,
     query,
@@ -26,12 +29,14 @@ import 'lex_property_generator.dart';
 }
 
 final class _LexXrpcQueryGenerator {
+  final GenContext ctx;
   final lex.NSID lexiconId;
   final String defName;
   final lex.LexXrpcQuery query;
   final List<String> mainVariants;
 
   _LexXrpcQueryGenerator(
+    this.ctx,
     this.lexiconId,
     this.defName,
     this.query,
@@ -49,6 +54,7 @@ final class _LexXrpcQueryGenerator {
     final parameters = query.parameters!;
 
     final properties = generateLexPropertiesFromLexXrpcParameters(
+      ctx,
       lexiconId,
       defName,
       parameters.properties,
@@ -84,6 +90,7 @@ final class _LexXrpcQueryGenerator {
     final object = output?.schema?.whenOrNull(object: (e) => e);
     if (object != null) {
       final properties = generateLexProperties(
+        ctx,
         lexiconId,
         defName,
         object.properties,