+++ packages/lex_gen/lib/src/services/object/lex_known_values.dart
@@ -4,6 +4,7 @@
 
 // Project imports:
 import '../../utils.dart';
+import '../gen_context.dart';
 import '../rule.dart' as rule;
 import 'lex_type.dart';
 
@@ -29,8 +30,14 @@ final class LexKnownValues extends GeneratableType {
   });
 
   @override
-  String getFilePath() {
-    return rule.getFilePath(lexiconId, defName, state, fieldName: fieldName);
+  String getFilePath(final GenContext ctx) {
+    return rule.getFilePath(
+      ctx,
+      lexiconId,
+      defName,
+      state,
+      fieldName: fieldName,
+    );
   }
 
   @override
@@ -51,7 +58,7 @@ final class LexKnownValues extends GeneratableType {
   }
 
   @override
-  String format() {
+  String format(final GenContext ctx) {
     final elements = values
         .map((e) {
           final buffer = StringBuffer();