+++ packages/lex_gen/lib/src/services/object/lex_output.dart
@@ -3,6 +3,7 @@
 // BSD-style license that can be found in the LICENSE file.
 
 // Project imports:
+import '../gen_context.dart';
 import 'lex_property.dart';
 import 'lex_type.dart';
 import 'utils.dart';
@@ -80,7 +81,7 @@ final class LexOutput extends GeneratableType {
   }
 
   @override
-  String format() {
+  String format(final GenContext ctx) {
     return renderFreezedDataClass(
       name: name,
       suffix: 'Output',